+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
 ## [20.2.3] - (not yet released)
 - Audit.EntityFramework: Fix to honor Override/Ignore/Format in entity properties for any mapper type (#584)
+- Audit.EntityFramework: Fix EntityType null in the ExplicitMapper when using sync methods (#585)
 
 ## [20.2.2] - 2023-03-14
 - Audit.EntityFramework / Audit.EntityFramework.Core: Low-Level interceptors support to use the `.AuditDataProvider` property from the `AuditDbContext` instance (#387)